move view cache to Format::Render::Cache
